Fix for getting the last message.
